Basic user management
Users can now edit their email address, first name, last name and
address on the profile page.
The profile "show" page always redirects to profile "edit".

Admins can register new users with a link in the top bar
The registration page is only accessible by admins. All the admin adds
is the email, the system then sends a reset link to the user so that
the user can set a new password.

Users can reset their password
There is now a link on the login page that the user can click to get
to the "reset password" page and request a reset link via email.