Added invitations and invitation mail
Admins can create invitations in sonata.
An email us automatically sent to the person that is invited.
The email includes the URL and the code

When a user registers, the invitation code is mandatory.
It must match with any unused invitation code.
The email of the invitation can differ from the email of the
registration.